face_generation/requirements.txt
numpy
pandas
matplotlib==2.0.2
scikit-learn
jupyter
notebook
tensorflow-gpu==1.1.0
tqdm==4.11.2
pillow
